recipes in cart for unauthorised users

+++ apps/api/views.py
@@ -75,24 +75,40 @@ class SubscriptionApi(APIView):
 
 class CartAPI(APIView):
     permission_classes = [AllowAny, ]
-    # TODO: make smth with anonymous users
+    resp_mesg = {
+        201: 'successfully created',
+        400: 'recipe already in shop list',
+        200: 'successfully deleted',
+    }
 
     def post(self, request, *args, **kwargs):
-
         recipe = Recipe.objects.get(slug=request.data.get('recipe_slug'))
-        _, created = CartItem.objects.get_or_create(
-            user=self.request.user,
-            recipe=recipe)
-        if created:
-            return Response({'status': 'successfully created'},
-                            status=status.HTTP_201_CREATED)
-        return Response({'status': 'follow instance already exists'},
-                        status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated:
+            _, created = CartItem.objects.get_or_create(
+                user=self.request.user,
+                recipe=recipe)
+            if created:
+                return Response({'status': self.resp_mesg[201]},
+                                status=status.HTTP_201_CREATED)
+            return Response({'status': self.resp_mesg[400]},
+                            status=status.HTTP_400_BAD_REQUEST)
+        if request.session.get('cart') is None:
+            request.session['cart'] = [recipe.pk, ]
+        else:
+            if recipe.pk in request.session['cart']:
+                return Response({'status': self.resp_mesg[400]},
+                                status=status.HTTP_400_BAD_REQUEST)
+            request.session['cart'].append(recipe.pk)
+            request.session.modified = True
+        return Response({'status': self.resp_mesg[201]},
+                        status=status.HTTP_201_CREATED)
 
     def delete(self, request, *args, **kwargs):
-        get_object_or_404(CartItem,
-                          user=self.request.user,
-                          recipe__slug=kwargs.get('recipe_slug'),
-                          ).delete()
-        return Response({'status': 'successfully deleted'},
+        if request.user.is_authenticated:
+            get_object_or_404(CartItem,
+                              user=self.request.user,
+                              recipe__slug=kwargs.get('recipe_slug'),
+                              ).delete()
+
+        return Response({'status': self.resp_mesg[200]},
                         status=status.HTTP_200_OK)
